backend/requirements.txt
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==4.2.0
pandas==2.0.3
openpyxl==3.1.2
icalendar==5.0.7
psycopg2-binary==2.9.6